requirement.txt
Jinja2>=3.1
PyMySQL>=1.0
PyYAML>=6.0
pyhive>=0.7
mcp>=1.0
